Removed command line fixture into a single file to avoid clashes in namespace.
