+++ Src/Tests/SolveTests.cs
@@ -117,23 +117,4 @@ namespace Tests
             Assert.True(SolvePartialModel("pm"));
         }
     }
-
-    public class CommandInterfaceFixture : IDisposable
-    {
-        private readonly CommandLineProgram.ConsoleChooser _chooser;
-        private readonly CommandLineProgram.ConsoleSink _sink;
-        public CommandInterface _ci { get; private set; }
-
-        public CommandInterfaceFixture()
-        {
-            _chooser = new CommandLineProgram.ConsoleChooser();
-            _sink = new CommandLineProgram.ConsoleSink();
-            _ci = new CommandInterface(_sink, _chooser);
-        }
-
-        public void Dispose()
-        {
-            _ci.Cancel();
-        }
-    }
 }
